Update return to pacientes controller

+++ src/main/java/com/backend/clinica/controller/PacienteController.java
@@ -3,6 +3,7 @@ package com.backend.clinica.controller;
 import com.backend.clinica.entity.PacienteModel;
 import com.backend.clinica.service.IService;
 import com.backend.clinica.service.PacienteService;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
@@ -17,27 +18,69 @@ public class PacienteController {
     }
 
     @GetMapping
-    public ArrayList<PacienteModel> getPacientes() {
-        return pacienteService.findAll();
+    public ResponseEntity<CustomResponse> getPacientes() {
+        ArrayList<PacienteModel> pacientes = pacienteService.findAll();
+        if (pacientes.isEmpty()) {
+            CustomResponse cr = new CustomResponse(false, "No hay pacientes en la base de datos", null);
+            return ResponseEntity.status(404).body(cr);
+        } else {
+            CustomResponse cr = new CustomResponse(true, "Pacientes encontrados", pacientes);
+            return ResponseEntity.status(200).body(cr);
+        }
     }
 
     @GetMapping("/{id}")
-    public PacienteModel getPaciente(@PathVariable("id") Integer id) {
-        return pacienteService.findById(id);
+    public ResponseEntity<CustomResponse> getPaciente(@PathVariable("id") Integer id) {
+        PacienteModel paciente = pacienteService.findById(id);
+
+        if (paciente == null) {
+            CustomResponse cr = new CustomResponse(false, "Paciente no encontrado", null);
+            return ResponseEntity.status(404).body(cr);
+        } else {
+            CustomResponse cr = new CustomResponse(true, "Paciente encontrado", paciente);
+            return ResponseEntity.status(302).body(cr);
+        }
     }
 
     @PostMapping
-    public PacienteModel createPaciente(@RequestBody PacienteModel pacienteModel) {
-        return pacienteService.create(pacienteModel);
+    public ResponseEntity<CustomResponse> createPaciente(@RequestBody PacienteModel pacienteModel) {
+        PacienteModel paciente = pacienteService.create(pacienteModel);
+
+        if (paciente == null) {
+            CustomResponse cr = new CustomResponse(false, "Error al crear el paciente", null);
+            return ResponseEntity.status(400).body(cr);
+        } else {
+            CustomResponse cr = new CustomResponse(true, "Paciente creado correctamente", paciente);
+            return ResponseEntity.status(201).body(cr);
+        }
     }
 
     @PutMapping("/{id}")
-    public PacienteModel updatePaciente(@RequestBody PacienteModel pacienteModel, @PathVariable("id") Integer id) {
-        return pacienteService.update(pacienteModel, id);
+    public ResponseEntity<CustomResponse> updatePaciente(@RequestBody PacienteModel pacienteModel, @PathVariable("id") Integer id) {
+        PacienteModel paciente = pacienteService.findById(id);
+
+        if (paciente == null) {
+            CustomResponse cr = new CustomResponse(false, "Paciente no encontrado", null);
+            return ResponseEntity.status(404).body(cr);
+        } else {
+            pacienteModel.setPacienteID(id);
+            PacienteModel pacienteUpdated = pacienteService.update(pacienteModel, id);
+            CustomResponse cr = new CustomResponse(true, "Paciente actualizado correctamente", pacienteUpdated);
+            return ResponseEntity.status(200).body(cr);
+        }
     }
 
     @DeleteMapping("/{id}")
-    public boolean deletePaciente(@PathVariable("id") Integer id) {
-        return pacienteService.delete(id);
+    public ResponseEntity<CustomResponse> deletePaciente(@PathVariable("id") Integer id) {
+        PacienteModel paciente = pacienteService.findById(id);
+
+        if (paciente == null) {
+            CustomResponse cr = new CustomResponse(false, "Paciente no encontrado", null);
+            return ResponseEntity.status(404).body(cr);
+        } else {
+            pacienteService.delete(id);
+            CustomResponse cr = new CustomResponse(true, "Paciente eliminado correctamente", null);
+            return ResponseEntity.status(200).body(cr);
+        }
     }
 }
